requirements_gradio.txt
gradio==3.50.2
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
pillow==10.0.1
numpy==1.24.3
pydantic==1.10.13